Fibonacci Evens Addition Challenge
Todays math challenge involved adding up all the even values of the Fib
sequence that are less than 4000000
Learned about iteration and keeping track of variables, like using temp
variables.
Learned about the challenges of the scopes of variables and how
important it is to keep track of them.

Output was 4613732